Add files via upload
I wrote this program based on a recurring problem in my last two algebra classes. It uses the formula for the relative maximum of a vertex of a parabola in order to calculate length and width of a fenced in area. The formula calculates in order to maximize area and assumes that one side is bordered by something like a river or a wall.